Add UITest helper class
